Creating the matrix and assigning the nodes
We've created a struct to save the matrix for the flowered game. we tried to make the size of struct at low as possible. we only defined a char variable to identify the color of index. (600 Bytes for the whole matrix in 5x5)
We will get answer if the game generated as 5x5. (Size didn't determined)
We are trying to make it better.